Got the quiz question recycler view to show up
The create quiz screen shows the title of the quiz, the topic spinner, and the first question, which has a question statement, 4 answer choices, and 4 checkboxes. The FAB is not in the right place though. Also the "No Students" now shows up when the user deletes the last student as well as when it starts off empty.